+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/sandbox/TwoDriverTeleOpMode.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.sandbox;
 
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
