+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/sandbox/bodhiMachineVision.java
@@ -19,7 +19,7 @@
  * SOFTWARE.
  */
 
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.sandbox;
 
 
 import java.lang.reflect.Array;
