+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/sandbox/machineVision.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.sandbox;
 
 import java.io.File;
 import java.util.Locale;
